SCREAM: reorganized some files locations
