+++ components/scream/src/control/atmosphere_driver.cpp
@@ -1,6 +1,6 @@
-#include "atmosphere_driver.hpp"
+#include "control/atmosphere_driver.hpp"
 
-#include "share/atmosphere_process_group.hpp"
+#include "share/atm_process/atmosphere_process_group.hpp"
 #include "share/scream_assert.hpp"
 #include "share/util/string_utils.hpp"
