+++ components/scream/src/control/tests/dummy_atm_proc.hpp
@@ -1,7 +1,7 @@
 #include <catch2/catch.hpp>
-#include "share/atmosphere_process.hpp"
-#include "share/scream_pack.hpp"
+#include "share/atm_process/atmosphere_process.hpp"
 #include "share/grid/user_provided_grids_manager.hpp"
+#include "share/scream_pack.hpp"
 #include "control/atmosphere_driver.hpp"
 
 namespace scream {
