+++ components/scream/src/dynamics/homme/atmosphere_dynamics.hpp
@@ -1,7 +1,7 @@
 #ifndef SCREAM_HOMME_DYNAMICS_HPP
 #define SCREAM_HOMME_DYNAMICS_HPP
 
-#include "share/atmosphere_process.hpp"
+#include "share/atm_process/atmosphere_process.hpp"
 #include "share/scream_parameter_list.hpp"
 
 #include <string>
