+++ components/scream/src/interface/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(ATM_SRC
   scream_cxx_f90_interface.cpp
+  surface_coupling.cpp
   atm_comp_mct.F90
   dead_mct_mod.F90
   dead_data_mod.F90
