+++ components/scream/src/physics/shoc/tests/shoc_stand_alone.cpp
@@ -1,5 +1,5 @@
 #include <catch2/catch.hpp>
-#include "share/atmosphere_process.hpp"
+#include "share/atm_process/atmosphere_process.hpp"
 #include "share/scream_pack.hpp"
 #include "share/grid/user_provided_grids_manager.hpp"
 #include "share/grid/se_grid.hpp"
