+++ components/scream/src/share/atm_process/atmosphere_process_group.cpp
@@ -1,4 +1,4 @@
-#include "atmosphere_process_group.hpp"
+#include "share/atm_process/atmosphere_process_group.hpp"
 
 #include "share/util/scream_std_utils.hpp"
 #include "share/util/string_utils.hpp"
