+++ components/scream/src/share/grid/remap/combo_remapper.hpp
@@ -1,7 +1,7 @@
 #ifndef SCREAM_COMBO_REMAPPER_HPP
 #define SCREAM_COMBO_REMAPPER_HPP
 
-#include "share/remap/abstract_remapper.hpp"
+#include "share/grid/remap/abstract_remapper.hpp"
 
 namespace scream
 {
