+++ components/scream/tests/scream_p3/p3_stand_alone.cpp
@@ -1,5 +1,5 @@
 #include <catch2/catch.hpp>
-#include "share/atmosphere_process.hpp"
+#include "share/atm_process/atmosphere_process.hpp"
 #include "share/scream_pack.hpp"
 #include "share/grid/user_provided_grids_manager.hpp"
 #include "share/grid/se_grid.hpp"
